backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-generativeai==0.8.5
requests==2.31.0
python-dateutil==2.8.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0 